pyproject.toml
[project]
name = "medical-paper-summarizer"
version = "0.1.0"
description = "Medical paper summarizer demo"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "numpy",
    "torch>=2.9.1",
    "transformers>=5.2.0",
    "requests>=2.32.5",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "faiss-cpu>=1.13.2",
    "openai>=2.24.0",
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
